requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
pynput>=1.7.6
numpy>=1.24.0
pyperclip>=1.8.2
pystray>=0.19.5
Pillow>=10.0.0
torchaudio
nvidia-cublas-cu12
nvidia-cudnn-cu12
